+++ tests/test_session.py
@@ -1,7 +1,8 @@
 import unittest
 
-from tests.dummy_objects import (DummyClient, DummyDatabase, dummy_config)
+from tests.dummy_objects import (DummyClient, DummyDatabase)
 
+from asciichan.config import curry_configuration
 from asciichan.session import (curry_io, handle)
 
 
@@ -16,7 +17,7 @@ class IOTest(unittest.TestCase):
 class ClientHandler(unittest.TestCase):
     def setUp(self):
         self.client = DummyClient(b"quit")
-        self.config_get = dummy_config
+        self.config_get = curry_configuration("inexistent.conf")
 
     def test_handler(self):
         self.assertFalse(handle(self.client, "0.0.0.0", self.config_get))