+++ tests/test_shell.py
@@ -2,6 +2,7 @@ import unittest
 
 from tests.dummy_objects import (DummyClient, DummyDatabase)
 
+from asciichan.config import curry_configuration
 from asciichan.shell import (box_boards, box_posts, box_thread, shell)
 
 
@@ -16,6 +17,6 @@ class ShellTest(unittest.TestCase):
     def test_quit(self):
         client = DummyClient("quit")
         database = DummyDatabase()
-        config = configparser.ConfigParser()
+        config_get = curry_configuration("inexistent.conf")
         self.assertFalse(shell(client.send, client.recv, "a", "user",
-                               database, config))
+                               database, config_get))